+++ src/main/java/lotto/domain/WinningNumber.java
@@ -1,12 +1,12 @@
 package lotto.domain;
 
-import lotto.constant.ValidationErrorMessage;
+import lotto.constant.ErrorMessage;
 
 import java.util.List;
 
 public class WinningNumber extends Lotto {
-    static final String INVALID_RANGE_ERROR = ValidationErrorMessage.INVALID_RANGE.getErrorMessage();
-    static final String DUPLICATE_BONUS_NUMBER_ERROR = ValidationErrorMessage.DUPLICATE_BONUS_NUMBER.getErrorMessage();
+    static final String INVALID_RANGE_ERROR = ErrorMessage.INVALID_RANGE.getErrorMessage();
+    static final String DUPLICATE_BONUS_NUMBER_ERROR = ErrorMessage.DUPLICATE_BONUS_NUMBER.getErrorMessage();
 
     private int bonusNumber;
